// apps/client/src/theme/colors.ts
export const colorPalette = {
    transparent: 'transparent',
    default: {
        primary: {
            light: '#242424', //module
            main: '#121212', //background
            dark: '#000000', //tab
        },
        secondary: {
            light: '#B3B3B3', //text
            main: '#FFFFFF', //text
            dark: '#1ED760', //buttons
        },
    },
};